drm/amdgpu: avoid memory allocation in the critical code path
When we run out of VMIDs we need to wait for some to become available.
Previously we were using a dma_fence_array for that, but this means that
we have to allocate memory.

Instead just wait for the first not signaled fence from the least recently
used VMID to signal. That is not as efficient since we end up in this
function multiple times again, but allocating memory can easily fail or
deadlock if we have to wait for memory to become available.

Signed-off-by: Christian König <christian.koenig@amd.com>
Signed-off-by: ryan.steed.usa <ryan.steed.usa@pm.me>